requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
fastapi
uvicorn
websockets
requests
joblib
geopy
schedule
earthengine-api
tensorflow
pyyaml
python-dotenv
openai
shap
